+++ src/Data/Functor/Deriving.hs
@@ -16,6 +16,8 @@ module Data.Functor.Deriving (
     , deriveFunctorOptions
     , makeFmap
     , makeFmapOptions
+    , makeReplace
+    , makeReplaceOptions
       -- * 'FFTOptions'
     , FFTOptions(..)
     , defaultFFTOptions